Inv validation service add, inv service methods add

+++ src/main/java/net/greatstart/model/ProjectDescription.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -26,7 +27,7 @@ public class ProjectDescription {
     private LocalDate addStart;
 
     @Column(name = "min_invest")
-    private Integer minInvestment;
+    private BigDecimal minInvestment;
 
     @Column(name = "verified")
     private Boolean isVerified;
